worker-server/requirements.txt
google-cloud-storage
google-cloud-bigquery[bqstorage,pandas]
pyarrow
functions-framework==2.1.2
pandas
numpy
scipy